Refaktor. Rename og flytt dataklasser til egne filer
